Can Complete and Remove items, items are sorted
Now that items are sorted, a new function is used to generate the ID when adding items to the list. Without this, the same ID was being given to multiple items, as the newest item wasn't automatically at the end of the list. New functions have also been added to remove items from the list and to Complete items.