Updated main page
The main page now has three columns, "To Do", "Doing", and "Done". Each column has it's own list of cards. Cards in each column have relevant buttons to move them to other columns. This required a change to the update function in app.py to act appropriately according to which button was pressed. There's a lot of repeated HTML here so I'm working if there's a more elegant way to display these columns, but for now this works just fine.